# modules/academy-api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Academy Management System API
  version: 1.0.0
  description: |
    REST API for Academy Management System. This API provides endpoints for managing students, batches, classes, mentors, and mentor sessions.
    
    The API supports:
    - Student registration and management
    - Batch creation and management
    - Class scheduling and management
    - Mentor management
    - Mentor session booking and management
    
    All create operations publish Kafka events for event-driven architecture.
  contact:
    name: Academy Development Team
    email: support@academy.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://academy-backend-dev.example.com
    description: Development environment
  - url: https://academy-backend.example.com
    description: Production environment
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/validate:
    post:
      summary: Validate JWT token
      operationId: validateToken
      tags:
        - Authentication
      description: Validate a JWT token and return validation result
      security: []  # Public endpoint, no authentication required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenValidationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/batches:
    get:
      summary: Get all batches (paginated)
      operationId: getAllBatches
      tags:
        - Batches
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageBatch'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      summary: Create a new batch
      operationId: createBatch
      tags:
        - Batches
      description: Create a batch and publish batch.created Kafka event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchInput'
      responses:
        '201':
          description: Batch created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/batches/{id}:
    get:
      summary: Get batch by ID
      operationId: getBatchById
      tags:
        - Batches
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      summary: Update batch
      operationId: updateBatch
      tags:
        - Batches
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchInput'
      responses:
        '200':
          description: Batch updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete batch
      operationId: deleteBatch
      tags:
        - Batches
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '204':
          description: Batch deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/batches/{id}/classes/{classId}:
    post:
      summary: Assign class to batch
      operationId: assignClassToBatch
      tags:
        - Batches
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Batch ID
        - name: classId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Class ID
      responses:
        '200':
          description: Class assigned to batch successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/classes:
    get:
      summary: Get all classes
      operationId: getAllClasses
      tags:
        - Classes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Class'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      summary: Create a new class
      operationId: createClass
      tags:
        - Classes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassInput'
      responses:
        '201':
          description: Class created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/classes/{id}:
    get:
      summary: Get class by ID
      operationId: getClassById
      tags:
        - Classes
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      summary: Update class
      operationId: updateClass
      tags:
        - Classes
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClassId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassInput'
      responses:
        '200':
          description: Class updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete class
      operationId: deleteClass
      tags:
        - Classes
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '204':
          description: Class deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/batch-types:
    get:
      summary: Get all batch types
      operationId: getAllBatchTypes
      tags:
        - Batch Types
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchType'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new batch type
      operationId: createBatchType
      tags:
        - Batch Types
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchTypeInput'
      responses:
        '201':
          description: Batch type created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/batch-types/{id}:
    get:
      summary: Get batch type by ID
      operationId: getBatchTypeById
      tags:
        - Batch Types
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchTypeId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchType'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update batch type
      operationId: updateBatchType
      tags:
        - Batch Types
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchTypeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchTypeInput'
      responses:
        '200':
          description: Batch type updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete batch type
      operationId: deleteBatchType
      tags:
        - Batch Types
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BatchTypeId'
      responses:
        '204':
          description: Batch type deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication. 
        Token should be sent in the Authorization header with "Bearer " prefix.
        Header name is configurable via jwt.header.name (default: Authorization).
        Header prefix is configurable via jwt.header.prefix (default: Bearer ).
        
        Example:
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
  parameters:
...
    BatchId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Batch ID
    ClassId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Class ID
...
    BatchTypeId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Batch type ID
  schemas:
...
    Batch:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Batch ID
        name:
          type: string
          description: Batch name
          example: FSD-2024-03
        startMonth:
          type: string
          format: date
          description: Start month of the batch
          example: "2024-03-01"
        currentInstructor:
          type: string
          description: Current instructor name
          example: John Instructor
        batchTypeId:
          type: integer
          format: int64
          description: Batch type ID
        batchTypeName:
          type: string
          description: Batch type name
          example: Full Stack Development
        classIds:
          type: array
          items:
            type: integer
            format: int64
          description: List of class IDs assigned to this batch
    BatchInput:
      type: object
      required:
        - name
        - startMonth
        - currentInstructor
        - batchTypeId
      properties:
        name:
          type: string
          description: Batch name
          example: FSD-2024-03
        startMonth:
          type: string
          format: date
          description: Start month of the batch
          example: "2024-03-01"
        currentInstructor:
          type: string
          description: Current instructor name
          example: John Instructor
        batchTypeId:
          type: integer
          format: int64
          description: Batch type ID
        classIds:
          type: array
          items:
            type: integer
            format: int64
          description: List of class IDs to assign to this batch
    Class:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Class ID
        name:
          type: string
          description: Class name
          example: Introduction to Java
        date:
          type: string
          format: date
          description: Class date
          example: "2024-03-15"
        time:
          type: string
          format: time
          description: Class time
          example: "10:00:00"
        instructor:
          type: string
          description: Instructor name
          example: Jane Smith
    ClassInput:
      type: object
      required:
        - name
        - date
        - time
        - instructor
      properties:
        name:
          type: string
          description: Class name
          example: Introduction to Java
        date:
          type: string
          format: date
          description: Class date
          example: "2024-03-15"
        time:
          type: string
          format: time
          description: Class time
          example: "10:00:00"
        instructor:
          type: string
          description: Instructor name
          example: Jane Smith
...
    PageBatch:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
    BatchType:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Batch type ID
        name:
          type: string
          description: Batch type name
          example: Full Stack Development
    BatchTypeInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Batch type name
          example: Full Stack Development
...
    TokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: JWT token to validate
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    TokenValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
          description: Whether the token is valid
          example: true
        username:
          type: string
          description: Username extracted from token (if valid)
          example: admin@academy.com
        error:
          type: string
          description: Error message (if validation failed)
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'